requirements.txt
PyYAML~=6.0.2
setuptools~=80.9.0
PyQt5~=5.15.11
matplotlib~=3.9.4
pandas~=2.2.3
numpy~=1.26.4
torch~=2.7.1
seaborn~=0.13.2
mlflow~=3.1.0
mlflow-skinny~=3.1.0
gymnasium~=1.1.1
gymnasium[classic-control]~=1.1.1
gymnasium[box2d]~=1.1.1
dm_control~=1.0.30
swig~=4.3.0
imageio~=2.37.0
opencv-python-headless~=4.10.0
